refactor(data): isolate persisted skill hydration
